test, expecters: Move loggedin expecter to console package
To test that ipv6 configured at cloud-init is really working we need to
remove the ipv6 configureation from LoggedIn expecter functions, but
that's massive to do it little by liettle the LoggedIn functions are
moved to "console" package and the functions at login.go are a wrapper
around this ones that also configure ipv6 addresses.
