+++ tests/console/BUILD.bazel
@@ -2,10 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["console.go"],
+    srcs = [
+        "console.go",
+        "login.go",
+    ],
     importpath = "kubevirt.io/kubevirt/tests/console",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/util/net/dns:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
