+++ tests/errors/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["panic.go"],
+    importpath = "kubevirt.io/kubevirt/tests/errors",
+    visibility = ["//visibility:public"],
+)
