+++ tests/errors/panic.go
@@ -1,0 +1,7 @@
+package errors
+
+func PanicOnError(err error) {
+	if err != nil {
+		panic(err)
+	}
+}
