+++ tests/libvmi/BUILD.bazel
@@ -6,13 +6,16 @@ go_library(
         "cloudinit.go",
         "factory.go",
         "network.go",
+        "pod.go",
         "storage.go",
         "vmi.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libvmi",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/controller:go_default_library",
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/containerdisk:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
