+++ tests/login.go
@@ -1,19 +1,17 @@
 package tests
 
 import (
-	"fmt"
-	"regexp"
 	"time"
 
-	v1 "kubevirt.io/client-go/api/v1"
-
 	expect "github.com/google/goexpect"
-	"google.golang.org/grpc/codes"
 
+	netutils "k8s.io/utils/net"
+
+	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/client-go/log"
-	"kubevirt.io/kubevirt/pkg/util/net/dns"
 	"kubevirt.io/kubevirt/tests/console"
+	"kubevirt.io/kubevirt/tests/libvmi"
 )
 
 // LoginToCirros call LoggedInFedoraExpecter but does not return the expecter
@@ -27,174 +25,98 @@ func LoginToCirros(vmi *v1.VirtualMachineInstance) error {
 // Alpine test VM
 func LoggedInCirrosExpecter(vmi *v1.VirtualMachineInstance) (expect.Expecter, error) {
 	virtClient, err := kubecli.GetKubevirtClient()
-	PanicOnError(err)
-	expecter, _, err := console.NewExpecter(virtClient, vmi, 10*time.Second)
+	expecter, err := console.LoggedInCirrosExpecter(vmi)
 	if err != nil {
 		return nil, err
 	}
-	hostName := dns.SanitizeHostname(vmi)
-
-	// Do not login, if we already logged in
-	err = expecter.Send("\n")
-	if err != nil {
-		expecter.Close()
-		return nil, err
-	}
-	_, _, err = expecter.Expect(regexp.MustCompile(`\$`), 10*time.Second)
-	if err == nil {
-		return expecter, nil
-	}
-
-	b := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: "login as 'cirros' user. default password: 'gocubsgo'. use 'sudo' for root."},
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: hostName + " login:"},
-		&expect.BSnd{S: "cirros\n"},
-		&expect.BExp{R: "Password:"},
-		&expect.BSnd{S: "gocubsgo\n"},
-		&expect.BExp{R: console.PromptExpression}})
-	resp, err := expecter.ExpectBatch(b, 180*time.Second)
-
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("Login: %v", resp)
-		expecter.Close()
-		return nil, err
-	}
-
-	err = configureConsole(expecter, true)
-	if err != nil {
-		expecter.Close()
-		return nil, err
-	}
-
 	return expecter, configureIPv6OnVMI(vmi, expecter, virtClient)
 }
 
 // LoggedInAlpineExpecter return prepared and ready to use console expecter for
 // Alpine test VM
 func LoggedInAlpineExpecter(vmi *v1.VirtualMachineInstance) (expect.Expecter, error) {
-	virtClient, err := kubecli.GetKubevirtClient()
-	PanicOnError(err)
-	expecter, _, err := console.NewExpecter(virtClient, vmi, 10*time.Second)
-	if err != nil {
-		return nil, err
-	}
-
-	b := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: "localhost login:"},
-		&expect.BSnd{S: "root\n"},
-		&expect.BExp{R: console.PromptExpression}})
-	res, err := expecter.ExpectBatch(b, 180*time.Second)
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("Login: %v", res)
-		expecter.Close()
-		return nil, err
-	}
-
-	err = configureConsole(expecter, false)
-	if err != nil {
-		expecter.Close()
-		return nil, err
-	}
-	return expecter, err
+	return console.LoggedInAlpineExpecter(vmi)
 }
 
 // LoggedInFedoraExpecter return prepared and ready to use console expecter for
 // Fedora test VM
 func LoggedInFedoraExpecter(vmi *v1.VirtualMachineInstance) (expect.Expecter, error) {
 	virtClient, err := kubecli.GetKubevirtClient()
-	PanicOnError(err)
-	expecter, _, err := console.NewExpecter(virtClient, vmi, 10*time.Second)
+	expecter, err := console.LoggedInFedoraExpecter(vmi)
 	if err != nil {
 		return nil, err
 	}
-	b := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BSnd{S: "\n"},
-		&expect.BCas{C: []expect.Caser{
-			&expect.Case{
-				// Using only "login: " would match things like "Last failed login: Tue Jun  9 22:25:30 UTC 2020 on ttyS0"
-				// and in case the VM's did not get hostname form DHCP server try the default hostname
-				R:  regexp.MustCompile(fmt.Sprintf(`(localhost|%s) login: `, vmi.Name)),
-				S:  "fedora\n",
-				T:  expect.Next(),
-				Rt: 10,
-			},
-			&expect.Case{
-				R:  regexp.MustCompile(`Password:`),
-				S:  "fedora\n",
-				T:  expect.Next(),
-				Rt: 10,
-			},
-			&expect.Case{
-				R:  regexp.MustCompile(`Login incorrect`),
-				T:  expect.LogContinue("Failed to log in", expect.NewStatus(codes.PermissionDenied, "login failed")),
-				Rt: 10,
-			},
-			&expect.Case{
-				R: regexp.MustCompile(fmt.Sprintf(`\[fedora@%s ~\]\$ `, vmi.Name)),
-				T: expect.OK(),
-			},
-		}},
-		&expect.BSnd{S: "sudo su\n"},
-		&expect.BExp{R: console.PromptExpression},
-	})
-	res, err := expecter.ExpectBatch(b, 3*time.Minute)
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("Login: %+v", res)
-		expecter.Close()
-		return expecter, err
-	}
-
-	err = configureConsole(expecter, false)
-	if err != nil {
-		expecter.Close()
-		return nil, err
-	}
-
 	return expecter, configureIPv6OnVMI(vmi, expecter, virtClient)
 }
 
 // ReLoggedInFedoraExpecter return prepared and ready to use console expecter for
 // Fedora test VM, when you are reconnecting (no login needed)
 func ReLoggedInFedoraExpecter(vmi *v1.VirtualMachineInstance, timeout int) (expect.Expecter, error) {
-	virtClient, err := kubecli.GetKubevirtClient()
-	PanicOnError(err)
-	expecter, _, err := console.NewExpecter(virtClient, vmi, 10*time.Second)
-	if err != nil {
-		return nil, err
-	}
-	b := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: console.PromptExpression}})
-	res, err := expecter.ExpectBatch(b, time.Duration(timeout)*time.Second)
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("Login: %+v", res)
-		expecter.Close()
-		return expecter, err
-	}
-	return expecter, err
+	return console.ReLoggedInFedoraExpecter(vmi, timeout)
 }
 
-func configureConsole(expecter expect.Expecter, shouldSudo bool) error {
-	sudoString := ""
-	if shouldSudo {
-		sudoString = "sudo "
+func configureIPv6OnVMI(vmi *v1.VirtualMachineInstance, expecter expect.Expecter, virtClient kubecli.KubevirtClient) error {
+	hasEth0Iface := func() bool {
+		hasNetEth0Batch := append([]expect.Batcher{
+			&expect.BSnd{S: "\n"},
+			&expect.BExp{R: console.PromptExpression},
+			&expect.BSnd{S: "ip a | grep -q eth0; echo $?\n"},
+			&expect.BExp{R: console.RetValue("0")}})
+		_, err := console.ExpectBatchWithValidatedSend(expecter, hasNetEth0Batch, 30*time.Second)
+		return err == nil
 	}
-	batch := append([]expect.Batcher{
-		&expect.BSnd{S: "stty cols 500 rows 500\n"},
+
+	hasGlobalIPv6 := func() bool {
+		hasGlobalIPv6Batch := append([]expect.Batcher{
+			&expect.BSnd{S: "\n"},
+			&expect.BExp{R: console.PromptExpression},
+			&expect.BSnd{S: "ip -6 address show dev eth0 scope global | grep -q inet6; echo $?\n"},
+			&expect.BExp{R: console.RetValue("0")}})
+		_, err := console.ExpectBatchWithValidatedSend(expecter, hasGlobalIPv6Batch, 30*time.Second)
+		return err == nil
+	}
+
+	clusterSupportsIpv6 := func() bool {
+		pod := libvmi.GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
+		for _, ip := range pod.Status.PodIPs {
+			if netutils.IsIPv6String(ip.IP) {
+				return true
+			}
+		}
+		return false
+	}
+
+	if !clusterSupportsIpv6() ||
+		(vmi.Spec.Domain.Devices.Interfaces == nil || len(vmi.Spec.Domain.Devices.Interfaces) == 0 || vmi.Spec.Domain.Devices.Interfaces[0].InterfaceBindingMethod.Masquerade == nil) ||
+		(vmi.Spec.Domain.Devices.AutoattachPodInterface != nil && !*vmi.Spec.Domain.Devices.AutoattachPodInterface) ||
+		(!hasEth0Iface() || hasGlobalIPv6()) {
+		return nil
+	}
+
+	addIPv6Address := append([]expect.Batcher{
+		&expect.BSnd{S: "\n"},
 		&expect.BExp{R: console.PromptExpression},
-		&expect.BSnd{S: "echo $?\n"},
-		&expect.BExp{R: console.RetValue("0")},
-		&expect.BSnd{S: fmt.Sprintf("%sdmesg -n 1\n", sudoString)},
-		&expect.BExp{R: console.PromptExpression},
-		&expect.BSnd{S: "echo $?\n"},
+		&expect.BSnd{S: "sudo ip -6 addr add fd10:0:2::2/120 dev eth0; echo $?\n"},
 		&expect.BExp{R: console.RetValue("0")}})
-	resp, err := expecter.ExpectBatch(batch, 30*time.Second)
+	resp, err := console.ExpectBatchWithValidatedSend(expecter, addIPv6Address, 30*time.Second)
 	if err != nil {
-		log.DefaultLogger().Infof("%v", resp)
+		log.DefaultLogger().Object(vmi).Infof("addIPv6Address failed: %v", resp)
+		expecter.Close()
+		return err
 	}
-	return err
+
+	time.Sleep(5 * time.Second)
+	addIPv6DefaultRoute := append([]expect.Batcher{
+		&expect.BSnd{S: "\n"},
+		&expect.BExp{R: console.PromptExpression},
+		&expect.BSnd{S: "sudo ip -6 route add default via fd10:0:2::1 src fd10:0:2::2; echo $?\n"},
+		&expect.BExp{R: console.RetValue("0")}})
+	resp, err = console.ExpectBatchWithValidatedSend(expecter, addIPv6DefaultRoute, 30*time.Second)
+	if err != nil {
+		log.DefaultLogger().Object(vmi).Infof("addIPv6DefaultRoute failed: %v", resp)
+		expecter.Close()
+		return err
+	}
+
+	return nil
 }
