+++ tests/utils.go
@@ -101,6 +101,7 @@ import (
 
 	"kubevirt.io/kubevirt/tests/console"
 	cd "kubevirt.io/kubevirt/tests/containerdisk"
+	testerrors "kubevirt.io/kubevirt/tests/errors"
 	"kubevirt.io/kubevirt/tests/flags"
 	"kubevirt.io/kubevirt/tests/libnet"
 	"kubevirt.io/kubevirt/tests/libvmi"
@@ -1551,25 +1552,7 @@ func GetRunningPodByLabel(label string, labelType string, namespace string, node
 }
 
 func GetPodByVirtualMachineInstance(vmi *v1.VirtualMachineInstance, namespace string) *k8sv1.Pod {
-	virtCli, err := kubecli.GetKubevirtClient()
-	PanicOnError(err)
-
-	pods, err := virtCli.CoreV1().Pods(namespace).List(metav1.ListOptions{})
-	PanicOnError(err)
-
-	var controlledPod *k8sv1.Pod
-	for _, pod := range pods.Items {
-		if controller.IsControlledBy(&pod, vmi) {
-			controlledPod = &pod
-			break
-		}
-	}
-
-	if controlledPod == nil {
-		PanicOnError(fmt.Errorf("no controlled pod was found for VMI"))
-	}
-
-	return controlledPod
+	return libvmi.GetPodByVirtualMachineInstance(vmi, namespace)
 }
 
 func GetComputeContainerOfPod(pod *k8sv1.Pod) *k8sv1.Container {
@@ -1733,9 +1716,7 @@ func createNamespaces() {
 }
 
 func PanicOnError(err error) {
-	if err != nil {
-		panic(err)
-	}
+	testerrors.PanicOnError(err)
 }
 
 func NewRandomDataVolumeWithRegistryImport(imageUrl, namespace string, accessMode k8sv1.PersistentVolumeAccessMode) *cdiv1.DataVolume {
@@ -2923,71 +2904,6 @@ func RenderPod(name string, cmd []string, args []string) *k8sv1.Pod {
 	return &pod
 }
 
-func configureIPv6OnVMI(vmi *v1.VirtualMachineInstance, expecter expect.Expecter, virtClient kubecli.KubevirtClient) error {
-	hasEth0Iface := func() bool {
-		hasNetEth0Batch := append([]expect.Batcher{
-			&expect.BSnd{S: "\n"},
-			&expect.BExp{R: console.PromptExpression},
-			&expect.BSnd{S: "ip a | grep -q eth0; echo $?\n"},
-			&expect.BExp{R: console.RetValue("0")}})
-		_, err := console.ExpectBatchWithValidatedSend(expecter, hasNetEth0Batch, 30*time.Second)
-		return err == nil
-	}
-
-	hasGlobalIPv6 := func() bool {
-		hasGlobalIPv6Batch := append([]expect.Batcher{
-			&expect.BSnd{S: "\n"},
-			&expect.BExp{R: console.PromptExpression},
-			&expect.BSnd{S: "ip -6 address show dev eth0 scope global | grep -q inet6; echo $?\n"},
-			&expect.BExp{R: console.RetValue("0")}})
-		_, err := console.ExpectBatchWithValidatedSend(expecter, hasGlobalIPv6Batch, 30*time.Second)
-		return err == nil
-	}
-
-	clusterSupportsIpv6 := func() bool {
-		pod := GetPodByVirtualMachineInstance(vmi, vmi.Namespace)
-		for _, ip := range pod.Status.PodIPs {
-			if netutils.IsIPv6String(ip.IP) {
-				return true
-			}
-		}
-		return false
-	}
-
-	if !clusterSupportsIpv6() ||
-		(vmi.Spec.Domain.Devices.Interfaces == nil || len(vmi.Spec.Domain.Devices.Interfaces) == 0 || vmi.Spec.Domain.Devices.Interfaces[0].InterfaceBindingMethod.Masquerade == nil) ||
-		(vmi.Spec.Domain.Devices.AutoattachPodInterface != nil && !*vmi.Spec.Domain.Devices.AutoattachPodInterface) ||
-		(!hasEth0Iface() || hasGlobalIPv6()) {
-		return nil
-	}
-
-	addIPv6Address := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: console.PromptExpression},
-		&expect.BSnd{S: "sudo ip -6 addr add fd10:0:2::2/120 dev eth0; echo $?\n"},
-		&expect.BExp{R: console.RetValue("0")}})
-	resp, err := console.ExpectBatchWithValidatedSend(expecter, addIPv6Address, 30*time.Second)
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("addIPv6Address failed: %v", resp)
-		expecter.Close()
-		return err
-	}
-
-	time.Sleep(5 * time.Second)
-	addIPv6DefaultRoute := append([]expect.Batcher{
-		&expect.BSnd{S: "\n"},
-		&expect.BExp{R: console.PromptExpression},
-		&expect.BSnd{S: "sudo ip -6 route add default via fd10:0:2::1 src fd10:0:2::2; echo $?\n"},
-		&expect.BExp{R: console.RetValue("0")}})
-	resp, err = console.ExpectBatchWithValidatedSend(expecter, addIPv6DefaultRoute, 30*time.Second)
-	if err != nil {
-		log.DefaultLogger().Object(vmi).Infof("addIPv6DefaultRoute failed: %v", resp)
-		expecter.Close()
-		return err
-	}
-
-	return nil
-}
 func NewVirtctlCommand(args ...string) *cobra.Command {
 	commandline := []string{}
 	master := flag.Lookup("master").Value
